Enable "Respond via SMS" option
Enable "Respond via SMS" option, whether or not it is SMS pref
subscription. Since user knows whether MT call is on SMS pref
subscription or not, he can choose to use this option or not.
This would help have a consistent behavior even when
"Ask First" is set as SMS pref subscription.

Change-Id: Idc666c151b8519a350ec56bf2ad756c21d43d223
CRs-Fixed: 972040